Added new factory and default serttings file.
